refactor(cli): move fs related classes from schema-migrations to cli package
