+++ packages/cli/package.json
@@ -24,6 +24,7 @@
     "@contember/config-loader": "workspace:*",
     "@contember/database": "workspace:*",
     "@contember/dic": "workspace:*",
+    "@contember/engine-common": "workspace:*",
     "@contember/engine-content-api": "workspace:*",
     "@contember/schema": "workspace:*",
     "@contember/schema-migrations": "workspace:*",
