+++ packages/schema-migrations/src/index.ts
@@ -1,11 +1,7 @@
 export * from './SchemaDiffer'
 export * from './Migration'
-export * from './MigrationsResolver'
-export * from './SchemaVersionBuilder'
 export * from './SchemaMigrator'
 export * from './SchemaDiffer'
-export * from './MigrationCreator'
 export * from './MigrationDescriber'
 export * from './modifications'
-export * from './MigrationFilesManager'
 export { MigrationVersionHelper } from '@contember/engine-common'
